include updated_on in the list of fields updated by update()
If we're going to use a helper method to cherry-pick specific fields we
want to update on an instance, we have to make sure we tell django it's
okay to update the updated_on field as well.

Includes unit test to verify that updated_on changes when an experiment
is edited.